pyproject.toml
[tool.poetry]
name = "nonebot-plugin-molar-mass"
version = "0.1.0"
description = "A tool to calculate molar mass for middle school students."
authors = ["kifuan <kifuan@foxmail.com>"]
readme = "README.md"
packages = [
    { include = "nonebot_plugin_molar_mass" }
]
include = ["elements.json"]


[tool.poetry.dependencies]
python = "^3.9"
rply = "^0.7.8"
nonebot2 = "^2.0.0rc2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
